refactor: apply modularity
